move SecuredAPIKeyRestriction to extension
